upgraded to angular 6 using angular-cli to create lib project.
