--- .npmignore
@@ -1,4 +1,0 @@
-node_modules/
-aot/
-yarn-error.log
-tools/
