--- config-library.js
@@ -1,3 +1,0 @@
-export const nameLibrary = "angular2-knob";
-export const PATH_SRC = "src/";
-export const PATH_DIST = "dist/";
