--- index.ts
@@ -1,1 +1,0 @@
-export {KnobModule} from './src/angular2-knob';
