+++ package.json
@@ -1,57 +1,60 @@
 {
-  "name": "angular2-knob",
-  "version": "1.1.12",
-  "description": "Angular directive for Knob component powered by d3.js v4 without jQuery dependencies",
-  "main": "dist/angular2-knob.umd.js",
+  "name": "ngx-knob",
+  "version": "0.0.0",
   "scripts": {
-    "copy": "node tools/copy-package.js && node tools/cleanup.js",
-    "bundle": "rimraf dist && rollup -c rollup.config.umd.js && rollup -c rollup.config.esm.js && tsc",
-    "postbundle": "node tools/removecode.js",
-    "build": "npm run bundle && npm run copy"
+    "ng": "ng",
+    "start": "ng serve",
+    "build": "ng build",
+    "test": "ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e"
   },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/xzegga/angular-knob"
-  },
-  "keywords": [
-    "Knob",
-    "Angular",
-    "d3v4",
-    "angular-knob"
+  "private": true,
+  "packageScope": "@xmlking",
+  "packages": [
+    "packages/knob"
   ],
-  "author": "Raul Escamilla",
-  "license": "MIT",
-  "bugs": {
-    "url": "https://github.com/xzegga/angular-knob/issues"
-  },
-  "homepage": "https://github.com/xzegga/angular-knob#readme",
-  "dependencies": {
-    "@angular/cli": "^6.0.3",
+  "devDependencies": {
+    "@angular-devkit/build-angular": "~0.7.0",
+    "@angular-devkit/build-ng-packagr": "~0.7.0",
+    "@angular/animations": "^6.1.0",
+    "@angular/cli": "~6.1.2",
+    "@angular/common": "^6.1.0",
+    "@angular/compiler": "^6.1.0",
+    "@angular/compiler-cli": "^6.1.0",
+    "@angular/core": "^6.1.0",
+    "@angular/forms": "^6.1.0",
+    "@angular/http": "^6.1.0",
+    "@angular/language-service": "^6.1.0",
+    "@angular/platform-browser": "^6.1.0",
+    "@angular/platform-browser-dynamic": "^6.1.0",
+    "@angular/router": "^6.1.0",
     "@types/d3": "^4.10.0",
     "@types/d3-selection": "^1.1.0",
     "@types/d3-selection-multi": "^1.0.6",
+    "@types/jasmine": "~2.8.6",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/node": "~10.7.1",
+    "codelyzer": "~4.4.4",
+    "core-js": "^2.5.4",
     "d3": "^4.13.0",
-    "d3-selection-multi": "^1.0.1"
-  },
-  "devDependencies": {
-    "@angular/compiler": "^4.3.6",
-    "@angular/compiler-cli": "^4.3.6",
-    "@angular/core": "^4.4.7",
-    "@types/d3": "^4.13.0",
-    "@types/d3-selection": "^1.3.0",
-    "@types/d3-selection-multi": "^1.0.6",
     "d3-selection-multi": "^1.0.1",
-    "del": "^3.0.0",
-    "node-sass": "^4.9.0",
-    "rimraf": "^2.6.2",
-    "rollup": "^0.49.2",
-    "rollup-plugin-angular": "^0.5.3",
-    "rollup-plugin-commonjs": "^9.1.3",
-    "rollup-plugin-node-resolve": "^3.3.0",
-    "rollup-plugin-typescript": "^0.8.1",
-    "rxjs": "^5.5.11",
-    "typescript": "^2.8.3",
-    "uglify-js": "^3.0.28",
-    "zone.js": "^0.8.17"
+    "jasmine-core": "~3.2.1",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~3.0.0",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "~2.0.0",
+    "karma-jasmine": "~1.1.1",
+    "karma-jasmine-html-reporter": "^0.2.2",
+    "ng-packagr": "^4.1.0",
+    "prettier": "^1.14.2",
+    "protractor": "~5.4.0",
+    "rxjs": "^6.0.0",
+    "ts-node": "~7.0.1",
+    "tsickle": ">=0.29.0",
+    "tslib": "^1.9.0",
+    "tslint": "~5.11.0",
+    "typescript": "~2.7.2",
+    "zone.js": "~0.8.26"
   }
 }
