+++ projects/knob/ng-package.prod.json
@@ -1,0 +1,7 @@
+{
+  "$schema": "../../node_modules/ng-packagr/ng-package.schema.json",
+  "dest": "../../dist/knob",
+  "lib": {
+    "entryFile": "src/public_api.ts"
+  }
+}
