+++ projects/knob/src/lib/knob.module.ts
@@ -1,0 +1,10 @@
+import { NgModule } from '@angular/core';
+import { KnobDirective } from './knob.directive';
+
+@NgModule({
+  imports: [
+  ],
+  declarations: [KnobDirective],
+  exports: [KnobDirective]
+})
+export class KnobModule { }
