+++ projects/knob/src/public_api.ts
@@ -1,0 +1,6 @@
+/*
+ * Public API Surface of knob
+ */
+
+export * from './lib/knob.directive';
+export * from './lib/knob.module';
