+++ projects/knob/tslint.json
@@ -1,0 +1,17 @@
+{
+    "extends": "../../tslint.json",
+    "rules": {
+        "directive-selector": [
+            true,
+            "attribute",
+            "ngx",
+            "camelCase"
+        ],
+        "component-selector": [
+            true,
+            "element",
+            "ngx",
+            "kebab-case"
+        ]
+    }
+}
