--- src/angular2-knob.module.ts
@@ -1,9 +1,0 @@
-import {NgModule} from '@angular/core';
-import {Ng2KnobDirective} from './angular2-knob.directive';
-
-@NgModule({
-    exports: [Ng2KnobDirective],
-    declarations: [Ng2KnobDirective]
-})
-export class KnobModule { }
-
