--- src/angular2-knob.ts
@@ -1,1 +1,0 @@
-export { KnobModule } from './angular2-knob.module'
