+++ src/app/app.component.html
@@ -1,0 +1,1 @@
+<div ngxKnob [value]="value" [options]="knOptions"></div>
