--- tools/copy-package.js
@@ -1,3 +1,0 @@
-const fs = require('fs');
-let resizable = fs.readFileSync('package.json').toString();
-fs.writeFileSync('dist/package.json', resizable);
